Major User Manual Updates
Major updates to the user manual, the note-taking feature has been completely rewritten. It has been changed to "Add Comments" as the new title is a comment, not a note. Changed the export section to match the new changes and included all export options. Significant changes in the main sections of the manual, several rewrites of the preferences menu and reading options to include new options.